+++ routes/web.php
@@ -47,6 +47,8 @@ if (config('greenlight.compatibility')) {
     });
 }
 
+Route::get('metrics', \App\Http\Controllers\MetricsController::class)->middleware(['enable_if_config:metrics.enabled']);
+
 if (! env('DISABLE_CATCHALL_ROUTES')) {
     Route::any('/{any}', [ApplicationController::class, 'index'])->where('any', '.*')->middleware(Spatie\Csp\AddCspHeaders::class);
 }